+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adding `FixedReverseHeap`.
 * Adding `Heap.nsmallest` & `Heap.nlargest`.
+* Adding `MultiSet.isSubset` & `MultiSet.isSuperset`.
 * Adding `#.top` to `MultiSet`.
 * Adding missing `Heap` types.
 * Renaming `FiniteStack` to `FixedStack`.